Fix negative temperatures read as large positive values (TMP112)
Cast the TMP112 register value to int16_t before right-shifting to
preserve two's complement sign extension. Without the cast,
sys_get_be16 returns uint16_t and negative temperatures (e.g. -1 C
represented as 0xFF70) are interpreted as ~255 C.

+++ app/src/app_machine_probe.c
@@ -143,7 +143,7 @@ static int tmp112_read(const struct device *dev, float *temperature)
 	}
 
 	if (temperature) {
-		*temperature = (sys_get_be16(&read_buf[0]) >> 4) * 0.0625f;
+		*temperature = ((int16_t)sys_get_be16(&read_buf[0]) >> 4) * 0.0625f;
 	}
 
 	return 0;
